Highlight series on click.
Move selected series to the top!
To do this we had to switch to the go.Scattergl so we could add the two
sets of points for selected and other manually and in order to control
the layering.

Selected group shows up in the same color as they have in the main plot.
All other points are made a ttle translucent and gray.